chore: move lightbox location
