+++ apps/mobile/src/components/ui/lightbox/ImageViewing/components/ImageDefaultHeader.tsx
