+++ apps/mobile/src/components/ui/lightbox/ImageViewing/components/ImageItem/ImageItem.tsx
