+++ apps/mobile/src/components/ui/lightbox/ImageViewing/index.tsx
