+++ apps/mobile/src/components/ui/lightbox/lightboxState.tsx
