+++ apps/mobile/src/main.tsx
@@ -17,7 +17,7 @@ import { enableFreeze } from "react-native-screens"
 import { App } from "./App"
 import { BottomTabProvider } from "./components/layouts/tabbar/BottomTabProvider"
 import { BottomTabs } from "./components/layouts/tabbar/BottomTabs"
-import { Lightbox } from "./components/lightbox/Lightbox"
+import { Lightbox } from "./components/ui/lightbox/Lightbox"
 import { initializeApp } from "./initialize"
 import { followApi } from "./lib/api-client"
 import { apiClient } from "./lib/api-fetch"
