+++ apps/mobile/src/modules/entry-list/templates/EntrySocialItem.tsx
@@ -12,13 +12,13 @@ import type { MeasuredDimensions } from "react-native-reanimated"
 import Animated, { measure, runOnJS, runOnUI, useAnimatedRef } from "react-native-reanimated"
 
 import { useActionLanguage, useGeneralSettingKey } from "@/src/atoms/settings/general"
-import type { LightboxImageSource } from "@/src/components/lightbox/ImageViewing/@types"
-import { useLightboxControls } from "@/src/components/lightbox/lightboxState"
 import { UserAvatar } from "@/src/components/ui/avatar/UserAvatar"
 import { RelativeDateTime } from "@/src/components/ui/datetime/RelativeDateTime"
 import { FeedIcon } from "@/src/components/ui/icon/feed-icon"
 import { Image } from "@/src/components/ui/image/Image"
 import { getAllSources } from "@/src/components/ui/image/utils"
+import type { LightboxImageSource } from "@/src/components/ui/lightbox/ImageViewing/@types"
+import { useLightboxControls } from "@/src/components/ui/lightbox/lightboxState"
 import { ItemPressableStyle } from "@/src/components/ui/pressable/enum"
 import { ItemPressable } from "@/src/components/ui/pressable/ItemPressable"
 import { NativePressable } from "@/src/components/ui/pressable/NativePressable"
